Simplify the object store code. All database-backed first-class objects (Actor, Location) now inherit from SerObject, which handles a lot of the nasty low-level stuff for database serialisation and member access.
The code for member access allows both foo.bar and foo['bar']
accessors, with None returned if the member is undefined.

We also move the SQL schema to using BYTEA instead of TEXT types for
the state field, which allows us to use pickle protocol 2 (binary,
compressed) for serialising the objects.

Finally, we implement an OnLoad class, which does most of the work for
classes that want a method to be called immediately after they and
their parent class have been loaded from the database. (This last
feature is not fully implemented yet).