Add some methods (stack_bottom, stack_top, add_overlay) to Location to handle more stack management functions.
Fix a bug in the property access code where the fall-through to the
underlying overlays wasn't being handled correctly.

Add a FIXME comment about incorrect assumptions in the code
(set_mapping) that sets up the mapping-functions for handling
directions. This will need to be fixed later. For now, it is not
called at all.