Change "actor_id" column in actor table to "id", to fit with SerObject code (and the design style of the rest of the DB schema).
Remove NOT NULL constraints on location (x,y,layer) index fields.

Add (x,y,layer) index fields on actors.

As a result of Sert's schema changes in r1630:1632, we will be storing
all actors in the same table. This necessitates some changes to the
Actor/Player code. The main by-id cache is moved up the hierarchy, and
is now part of Actor. The by-name cache remains in Player, as only
Players have names.

Also add position indexing for all Actors, make a position a required
field in an Actor's constructor, and tweak the Player constructor to
handle it, too (with a note to fill in some code to make it alignment
dependent).