Exception handling and logging infrastructure.
Move the main debug log file to "debug"
Create a new log target for exceptions

Wrap up the whole of the main code in a try-except block, and handle
exceptions by dumping them to the wor-exceptions log file for later
perusal.

Also add a sketch of some to-database logging for game events.