Rework demand-loading of properties.
Using __getattr__ to do demand-loading of object properties is
problematic, as it only gets called if the attribute isn't found in
the object's dictionary, or in the class hierarchy. This means that if
a class has a default class property, any objects of that class (or
its subclasses) will not get their actual property loaded, as it finds
the attribute in the class first, and never calls __getattr__.

We fix this by taking complete control, with __getattribute__. We pass
properties beginning with _ to the default implementation
unconditionally (as they're not in the DB anyway). Anything else, we
look in the database first, then call the default implementation. We
also keep track of the failed attempts at loading from the database,
too, for performance reasons. Any request which has previously failed
is not tested in the DB.