Generalise the image-request filename processing.
Earlier versions of the on-demand image service required filenames in
strict forms. This update relaxes those forms and generalises them, so
that requested files can have any order of component tags, and the
filenames are canonicalised internally. This paves the way for rather
simpler handling of all the funny edge conditions.