Fix movement actions
Python binds variables in lambdas and functions by reference, so that
the value of the variable is set at execution time, not definition
time. This meant that the loop generating the movement actions was
creating up to six actions that all did what the last one did.

Use functools.partial, and a nasty little parameter-count hack
(Util.no_d()) to get around this behaviour.