Add information to help debug a particular (common) failure case.
We can panic (with a self-generated AssertionError) if we try to pickle
an object that contains a reference to a SerObject. The failure mode is
particularly obscure, and hard to track down by direct means. This patch
adds more comments and a message to help the unsuspecting programmer
that stumbles into this particular can of worms.